Plugin handling fixes.
In the case where the command prefix is found but no command matched,
the message would fall through to matchers, but the first word was
already removed as the expected command name.  Added an Args field to
Messages for commands and leave Text intact for matchers.

Added missing types to outgoing plugin messages.